Merged PR 135407: [MPT-7785] Use non-static dictionary for cache
I don't think this change should break anything, as the `EntityMapCache` is registered as singleton anyway.

The reasoning for this change: we had an issue with this in component tests for Notifications module. What we do there is set up two test servers - one for API and one for worker. They do however share the app domain, so they do share statics. This occurred to be a problem in some cases - for example, there is a test that starts with a worker action that produces audit events, which executes mapping logic and caches it. Later on, the API action reuses that map cache and is not able to translate the query. The same query is translatable if we, for example, remove the producer (so worker doesn't cache anything), or if the test is executed in a suite where some other action is done first. This makes it unstable and faulty. If no objections, I'd like to remove that static.